refactor: split better_player_localizations.dart into separate classes per file
