+++ packages/better_player/lib/src/controls/better_player_cupertino_localizations.dart
@@ -1,0 +1,14 @@
+import 'package:better_player/src/configuration/player_translations.dart';
+import 'package:cupertino_ui/cupertino_ui.dart';
+
+class BetterPlayerCupertinoLocalizations extends DefaultCupertinoLocalizations {
+  const BetterPlayerCupertinoLocalizations(this._translations);
+
+  final PlayerTranslations _translations;
+
+  @override
+  String get cancelButtonLabel => _translations.overflowMenuCancelLabel;
+
+  @override
+  String get modalBarrierDismissLabel => _translations.overflowMenuScrimLabel;
+}
